+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Mapbox welcomes participation and contributions from everyone.
 * Fix bearing value is fluctuating between initial value and correct value during a rotation gesture.
 * Allows animation during any ongoing gestures.
 * Sync map size to the size of the metal view.
+* Fix missing feature properties for `nil`/`null` values.
 
 ## 10.12.0 - March 22, 2023